teste

+++ app/Http/Controllers/VehicleController.php
@@ -214,13 +214,22 @@ class VehicleController extends Controller
 
     public function deleteSelected(Request $request)
     {
-        $selected_ids = explode(',', $request->selected_ids);
 
-        if (!empty($selected_ids)) {
-            Vehicle::whereIn('id', $selected_ids)->delete();
-            return redirect()->route('vehicles.index');
+        if ($request->has('selected_ids')) {
+
+            $selected_ids = json_decode($request->selected_ids);
+
+
+            if (!empty($selected_ids)) {
+
+                Vehicle::whereIn('id', $selected_ids)->delete();
+            }
         }
+
+
+        return redirect()->route('vehicles.index');
     }
+
 
 
     public function showVehicles()
